Fixed conflict problem with results file name and log file name
After the simulation has been started, I figured out that there is a conflict between the results file name and log file name. The results will be overwritten if two simulations start at the same day but with different GROMACS data sets. Therefore, I added the file name of the GROMACS simulation data file in the results file name and log file name. The random sequence of atoms that was used in the first simulation step is also copied from the server to use the same seqeunce for the other simulations of PLPC.